droplet-1-clicks/lamp: update lamp to work with dbaas
This set of changes should allow lamp to be used with
predeploy/dbaas. It will look for the credentials 
that predeploy writes to disk, and if they’re there
this means the user has requested a dbaas instance.
This will also ensure the local mysql instance on
the droplet is turned off, and it will also ping
the dbaas instance until it is ready.